@@ -4,7 +4,17 @@
 	<groupId>CoreSpr</groupId>
 	<artifactId>CoreSpr</artifactId>
 	<version>1.0</version>
+	<properties>
+		<spring.version>6.1.3</spring.version>
+		<spring.int.version>6.2.1</spring.int.version>
+		<jdk.version>21</jdk.version>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.target>21</maven.compiler.target>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+	</properties>
+
 	<dependencies>
+		<!-- Spring Framework 6 -->
 		<dependency>
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-core</artifactId>
