@@ -1,9 +1,9 @@
 package app.util;
 
 
-import javax.jms.ConnectionFactory;
-import javax.jms.Destination;
-import javax.jms.JMSException;
+import jakarta.jms.ConnectionFactory;
+import jakarta.jms.Destination;
+import jakarta.jms.JMSException;
 
 import org.springframework.jms.core.JmsTemplate;
 
